KUDU-2291 (part 1): allow collecting a thread's stack without immediate symbolization
For the /stacks page we would like to be able to grab the stack from a
bunch of threads as quickly as possible to eliminate "skid" between the
collection times. Symbolization of stacks is much slower than
collection, so this patch adds a new API to collect the remote thread
stack without symbolizing it immediately.

This adds a simple benchmark for remote thread stack collection. The
initial results are not that different between symbolized and
unsymbolized because the stack collection performance itself is poor.
This will be improved in a later patch in this series.

I0207 19:33:20.960477 30799 debug-util-test.cc:175] Throughput: 99 dumps/second (symbolize=0)
I0207 19:33:21.963946 30799 debug-util-test.cc:175] Throughput: 84 dumps/second (symbolize=1)

Change-Id: I0a14fbb7d0cafcd7b966dac507e71bae32e2cc1a
Reviewed-on: http://gerrit.cloudera.org:8080/9252
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>