tools: add more detail about unsafe_change_config
Given the tooling docs pages are auto-generated from the descriptions
provided by the tools directly, it'd be helpful to start describing when
to use each tool, in addition to the action itself.

In particular, in the docs page, when searching for the word "evict",
it'd be nice to be pointed to the `unsafe_change_config` tool.

Change-Id: If389f9fec4926b798c9885dac17b4de136644239
Reviewed-on: http://gerrit.cloudera.org:8080/9278
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins