[Spark] Simplifiy creating range partitioned tables
In order to create a range partitioned table a user
needs to create PartialRows representing the
lower and upper bounds. In the Spark API
the Kudu Schema isn’t made available which
could make matching the expected schema
tricky.

This small change makes retrieving the generated
Kudu schema possible and also allows for table
creation with a Kudu schema directly.

Change-Id: Ied45b90b38d9e123868c4b0430dc52c888f033b3
Reviewed-on: http://gerrit.cloudera.org:8080/9379
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Dan Burkert <danburkert@apache.org>