[itests] fix flake in raft_consensus_nonvoter-itest
Fixed flake in RaftConsensusNonVoterITest::RestartClusterWithNonVoter
scenario in the case when a newly added replica becomes a leader.

The flake was most prominent with ASAN builds.  Below are links to the
dist_test results of ASAN builds, with and without the fix:

before the fix (50 of 256 failed):
  http://dist-test.cloudera.org/job?job_id=aserbin.1519227435.49089

after the fix (0 of 256 failed):
  http://dist-test.cloudera.org/job?job_id=aserbin.1519227218.43636

Change-Id: I0b27eec4e9a4727b96689d4e9592b1f1684ebf54
Reviewed-on: http://gerrit.cloudera.org:8080/9380
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>