rpc: prevent death test from leaving behind test files
I noticed this while running tests locally, though I don't know why it
wasn't caught in precommit, as build-and-test.sh is supposed to look for
and flag files "left behind" by tests.

In any case, the issue is that a "threadsafe" death test forks and reruns
the test, which generates a new test directory and creates a new set of
test files. Then, the fork encounters a CHECK and crashes before cleaning up
these files.

Rather than fix this by making the death test a "fast" test, I opted to
refactor the associated methods to properly return errors, at which point
the test can evaluate the returned error.

Change-Id: I235400769a85d490039df07c3c084901534bf4e1
Reviewed-on: http://gerrit.cloudera.org:8080/9454
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>