build: use thin static libraries on Linux
Thin static libraries just contain pointers to .o files rather than copying all
of the data. This reduces the amount of IO during the build as well as the required
disk space.

For example, comparing a clean build ('ninja clean kudu') before and after this
change, we can see that the build is about 40% faster and about 200x
less disk space in the lib/ directory.

Thick (original):
  real    0m24.784s
  user    0m58.643s
  sys     0m37.701s

  todd@ve0518:/data/1/todd/kudu$ du -sm build/release/lib/
  594     build/release/lib/

Thin:
  real    0m17.422s
  user    0m56.092s
  sys     0m34.131s
  todd@ve0518:/data/1/todd/kudu$ du -sm build/release/lib/
  3       build/release/lib/

For comparison, a release build with -DKUDU_LINK=dynamic:

  real    0m24.008s
  user    2m3.212s
  sys     1m5.243s
  todd@ve0518:/data/1/todd/kudu/build/release-shared$ du -sm lib/
  256     lib/

Of course the dynamic build has a much smaller resulting 'kudu' binary, and if
we are compiling many binaries (eg all the tests), the dynamic build is still
faster and smaller than a thin-static build.

Change-Id: If662cea380e06eaddf45e52617e38e55e4613773
Reviewed-on: http://gerrit.cloudera.org:8080/9472
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins