[consensus] fix on RequestForPeer clean-up
The consensus queue can switch to the non-leader mode when exiting
from the RequestForPeer() method scope, so it's necessary to check
for that condition in the scoped clean-up object before updating
information on the peer's health.

The issue was found while running the raft_consensus_stress-itest
for the 3-2-3 re-replication scheme, so I don't think any additional
test is needed to cover this change.

Change-Id: I71b3b47bbf9b5011bd11534287f6ae8f760addd5
Reviewed-on: http://gerrit.cloudera.org:8080/9487
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>