build: fix ar flags to avoid a warning on newer OSes
726514abfe309b0b88a5d868129f494c379af954 enabled thin linking by
modifying the flags passed to 'ar' to 'cruT'. This caused the following
warning on my Ubuntu 16 box:

ar: `u' modifier ignored since `D' is the default (see `U')

I checked the CMake source and found that the default 'ar' command line
is:

  set(CMAKE_C_ARCHIVE_CREATE "<CMAKE_AR> qc <TARGET> <LINK_FLAGS> <OBJECTS>")

So, I think it's more appropriate to use 'qcT' instead of 'cruT' ('T' is
the flag which enables the thin libraries). I verified that this still
produces thin libraries and the warnings are gone.

Change-Id: Icef2ba9336196e13cbc43508720874442c3b4460
Reviewed-on: http://gerrit.cloudera.org:8080/9486
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins