[tools] KUDU-2331: Allow move tool to work when uninvolved tserver is down
The move tool requires a clean ksck for the tablet it's acting on
for safety reasons, but the check for this failed if a tablet server
was down, even if the tablet server didn't host a replica for the
tablet and wasn't the destination server. This fixes the move
command so a down tserver uninvolved in the move won't prevent the
move.

Change-Id: Id1fc60233f4f50f478da7ccb104e37df3067400c
Reviewed-on: http://gerrit.cloudera.org:8080/9510
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-on: http://gerrit.cloudera.org:8080/9791
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>