Fix RUN_FLAKY_ONLY build
Now that our tests have shard numbers appended (foo-test.<shard>)
according to ctest, the RUN_FLAKY_ONLY build regex was no longer
applying properly. It was converting the flaky test 'foo-test' to the
regex '^foo-test$' which would not match any of the shards.

This simply removes the '$' from the end of the regex. I attempted to
instead add a '(.\d+)?' to the regex, but ctest has a hard limit on the
number of '()' pairs in a regex so it was failing to compile the regex.
It's also not possible to pass multiple separate regexes on the command
line. Given that we never have test names where one is a prefix of
another, I think this is probably fine.

Change-Id: I3574bb88c2f3c465e3c659b80b122c1109053648
Reviewed-on: http://gerrit.cloudera.org:8080/9533
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins