[master] add feature flag for 3-4-3 scheme
This changelist addresses the case of misconfiguration of the
replica management scheme between masters and tablet servers
in the case when a master of version < 1.7 happen to work with
tablet servers of version >= 1.7.

Also, added a new run-time flag --heartbeat_incompatibility_is_fatal
to control whether the replica management scheme misconfiguration
should be treated as a fatal error for a tablet server.

This is a follow-up for 1769eed53ee2c21a88a766cb72bf8c9ae622099d.

Change-Id: I0e34a6fffc375720b26c2641e88057975d9190b9
Reviewed-on: http://gerrit.cloudera.org:8080/9565
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>